example_agents/googleadk_academic_research_agent/pyproject.toml
[project]
name = "academic-research"
version = "0.1.0"
description = "AI-driven agent designed to facilitate the exploration of the academic landscape"
authors = [{ name = "Antonio Gulli", email = "gulli@google.com" }]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "google-adk>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.7",
    "google-adk[eval]>=1.0.0",
    "google-cloud-aiplatform[adk,agent-engines,evaluation]>=1.93.0",
    "nest-asyncio>=1.6.0",
    "pandas>=2.2.3",
    "tabulate>=0.9.0",
    "agent-starter-pack>=0.14.1",
    "ruff>=0.4.6",
    "mypy>=1.15.0",
    "codespell>=2.2.0",
    "types-pyyaml>=6.0.12.20240917",
    "types-requests>=2.32.0.20240914",
]

deployment = [
    "absl-py>=2.2.1",
]

[tool.ruff]
extend = "../../../pyproject.toml"

[tool.ruff.lint]
# Ignores function complication rule
ignore = ["C901"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]


[tool.ruff.lint.isort]
known-first-party = ["academic_research"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.agent-starter-pack]
example_question = "Analyze the Attention is All You Need paper"

[tool.agent-starter-pack.settings]
agent_directory = "academic_research"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"